Fix race condition

+++ R/crew_store_local.R
@@ -67,8 +67,6 @@ crew_store_local <- R6::R6Class(
         wait = private$wait,
         message = "store timeout moving local file after writing"
       )
-      
-      write(paste("ok writing file", path, Sys.getpid()), "~/Desktop/log.txt", append = TRUE)
     },
     exists_task = function(direction, worker) {
       crew_true(direction, is.character(.), !anyNA(.), length(.) == 1L)
@@ -83,15 +81,22 @@ crew_store_local <- R6::R6Class(
       crew_true(direction, is.character(.), !anyNA(.), length(.) == 1L)
       crew_true(worker, is.character(.), !anyNA(.), length(.) == 1L)
       path <- file.path(private$root, direction, worker)
-     path_temp <- file.path(private$root, "temp", crew_name())
-     crew_wait(
-       fun = function(path, path_temp) file.rename(from = path, to = path_temp),
-       args = list(path = path, path_temp = path_temp),
-       timeout = private$timeout,
-       wait = private$wait,
-       message = "store timeout moving local file for deletion"
-     )
-     unlink(path_temp, recursive = TRUE, force = TRUE)
+      if (!any(file.exists(path))) return()
+      path_temp <- file.path(private$root, "temp", crew_name())
+      crew_wait(
+        fun = function(path, path_temp) file.rename(from = path, to = path_temp),
+        args = list(path = path, path_temp = path_temp),
+        timeout = private$timeout,
+        wait = private$wait,
+        message = "store timeout moving local file for deletion"
+      )
+      unlink(path_temp, recursive = TRUE, force = TRUE)
+    },
+    log_worker = function(worker) {
+      crew_true(worker, is.character(.), !anyNA(.), length(.) == 1L)
+      path <- paste0(file.path(private$root, "log", worker), ".txt")
+      dir_create(dirname(path))
+      path.expand(path)
     },
     marshal_store = function() {
       expr <- substitute(
